requirements.txt
scapy==2.5.0
flask==3.0.0
flask-socketio==5.3.6
python-nmap==0.7.1
psutil==5.9.8
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
rich==13.7.0
click==8.1.7
sqlalchemy==2.0.25
python-dotenv==1.0.1
requests==2.31.0
colorama==0.4.6
networkx==3.2.1
matplotlib==3.8.2
plotly==5.18.0
